chore(integrations): reorganize files
