+++ src/modules/auth/auth-service.ts
@@ -1,6 +1,7 @@
 import { Repository } from 'typeorm';
 import crypto from 'crypto';
-import { ApiKey, IntegrationUser } from './entities';
+import { ApiKey } from './entities';
+import { IntegrationUser } from './integration/entities';
 import dataSource from '../../database';
 import { Audio, LightsController, Screen } from '../root/entities';
