+++ src/modules/auth/entities/index.ts
@@ -1,9 +1,7 @@
 import Session from './session';
 import ApiKey from './api-key';
-import IntegrationUser from './integration-user';
 
 export { default as Session } from './session';
 export { default as ApiKey } from './api-key';
-export { default as IntegrationUser } from './integration-user';
 
-export const Entities = [Session, IntegrationUser, ApiKey];
+export const Entities = [Session, ApiKey];
