+++ src/modules/auth/integration/entities/index.ts
@@ -1,0 +1,5 @@
+import IntegrationUser from './integration-user';
+
+export { default as IntegrationUser } from './integration-user';
+
+export const Entities = [IntegrationUser];
