+++ src/modules/auth/integration/index.ts
@@ -1,0 +1,3 @@
+export { default as IntegrationEndpointManager } from './integration-endpoint-manager';
+export { IntegrationUserController } from './integration-user-controller';
+export { default as IntegrationUserService } from './integration-user-service';
